Transfered  implementations of models into separate namespace.
