+++ NewsMaking.DesktopApp.ViewModels/NewsMaking.DesktopApp.ViewModels.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
+﻿<Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
 
   <PropertyGroup>
     <TargetFramework>net6.0-windows</TargetFramework>
